+++ release-notes.md
@@ -1,2 +1,3 @@
-1. Enhance the DBLP scaper for ICLR 2016.
-2. Fixed some bugs.
+1. Enhance the DBLP scaper for some workshop papers.
+2. Fixed the bibtex output for special char.
+3. Fixed some bugs.